Initial python module and packaging work
This commit introduces a Python module that uses ctypes to load any of
the four highctidh shared library files. Once installed it may be used
by importing the module, and setting the field size:

 import highctidh
 ctidh = highctidh.ctidh(512)

This commit additionally adds 29 tests which cover the highctidh API.
Run the module unit tests from the root of the source:

  python3 -m unittest -v

Optional Debian packaging is now available using stdeb. Build a Debian
package that also includes the shared library files is easy:

 pip install stdeb
 python3 setup.py bdist_deb
 sudo ldconfig