Portable libhighctidh enhancements
This commit merges initial portability work for libhighctidh on the
GNU/Linux platform when built with gcc-10. libhighctidh is now
functional with GNU/Linux on aarch64/arm64, loongarch64/Loongson,
ppc64le/POWER9, ppc64/POWER8, riscv64, s390x, sparc64, and x86_64.

The portable backend was initially generated by fiat-crypto, and then
some performance improvements were made to the generated C code.
libhighctidh remains backward compatible with the original reference
implementation. The original assembly implementation may be used
interchangeably by rebuilding the library (`make libhighctidh.so
HIGHCTIDH_PORTABLE=0`).  The Makefile was enhanced to add improved clean
and test targets.  All field sizes (511, 512, 1024, 2048) are supported
across all architectures. The primes chosen for the fiat-crypto
synthesis are documented in `PRIMES.md` as are the commands to reproduce
the original synthesis with fiat-crypto. Additionally included is a
portable public domain integer sorting implementation taken from
djbsort-20190516 which is used when the portable backend is selected.
The portable backend interoperates across big and little endian systems.
Many tests were added in both C and in Python. Basic benchmarking is
shown in `BENCHMARKING.md` and the data was gathered with
`highctidh-simple-benchmark.py` which is run as part of the `make test`
target. Python 3.9.2 is now required, though Python 3.10.0 is preferred.

Authors of this commit are:
Johan Kjaer
Jacob Appelbaum

This work is dedicated in the public domain by both authors.